Rename BalanceGraph to LineGraph
