+++ src/renderer/components/LineGraph/linegraph.scss
