+++ src/renderer/style/index.scss
@@ -19,7 +19,7 @@
 @import '../components/Validator/validatorsimple.scss';
 @import '../components/Cards/cards.scss';
 @import '../components/Loading/loading.scss';
-@import '../components/BalanceGraph/balancegraph.scss';
+@import '../components/LineGraph/linegraph.scss';
 @import '../components/ErrorFrequencyGraph/errorfrequencygraph.scss';
 @import '../components/ResponseCodeError/responsecodeerror.scss';
 @import '../components/Prompt/inputprompt.scss';
